Add argument to method translator in configurator.
It's now possible to use CLI argument in a configuration file.
Argument with no value required are translated with these rules :
1) Remove trailing - ;
2) Substitute /-(.)/ with \u\1 ;
So, '--loop' argument is translated by configurator to method 'loop()'.
Argument with at least one value required are translated with these
rules :
1) Remove trailing - ;
2) Substitute /-(.)/ with \u\1 ;
3) Add 'set' before the result ;
So '--bootstrap-file' argument is translated by configurator to method
'setBootstrapFile()', and -bf argument is translated to method 'setBf()'.