atLeastOnce() is now called by default in mock asserter.
In a mock asserter, if a call assertion is defined and there is no call
to once(), exactly(), atLeastOnce() or never(), atLeastOnce() is
automaticaly used to validate the assertion.
So, it's now possible to write $this->mock($mock)->call('foo') because
it's equivalent to $this->mock($mock)->call('foo')->atLeastOnce().