use current working dir as base dir to search for .atoum files
on a shared installation of atoum, for example if there is a symlink to
atoum in the ~/bin directory, atoum will search for configuration file
from that dir.

for example, in that case :

/
|-- home
   |-- user
      | -- bin
          | -- atoum -> pathToAtoum/bin/atoum

If we call our tests from

/home/user/Projects/myProject

atoum will lookup for .atoum files in :
* /
* /user
* /user/bin

instead of searching in

* /
* /user
* /user/Projects
* /user/Projects/myProject

This works when atoum and the tested Project are installed in the same
directory but not when we call it from another dir.

Thoses changes make atoum search for .atoum files from the current
working directy instead of the atoum dir.