Add file path and line number for each fail in TAP report (#171)
Each test line considered as a fail (i.e 'not ok') now has one more line in
the diagnostic section of the report : /path/to/test/file.php:666
'/path/to/test/file.php' being the path to the failed test fail and '666' the line number.

Skipped tests are not considered as failure so they don't display these
informations.

This diagnostic line will always be the last of the section so parsing would
be easy.