Add phpClass::isFinal() method and the property phpClass::isFinal
Currently, atoum does not provide any assertion about final status of a
class.
And even if using final on a class seems to be a bad practice, it's not
always the case, so having an assertion about that is interesting.
To use phpClass::isFinal() and the short version phpClass:isFinal (aka
syntaxic sugar), just do:

<?php

$this->testedClass->isFinal(); // or
$this->testedClass->isFinal; // strictly equivalent to the line above.