Compare nested objects with object::isEqualTo()
This is a POC of a strategy to compare deeply nested objects. PHP throws
fatal errors when the nesting level is too high so we can't catch it and
fallback. We have to always apply this strategy.

Basically, how this works:

* we try to `json_encode` objects with a limited depth: if this works we can assume that their is no nesting and the `==` will work.
* if the `json_encode` fails and the value is an object, we cast it to an array containing all properties (public, protected and private) we then clean the array removing closures for further serialization (we replace each closure instance by its spl_hash)
* we then walk the array to remove recursion by replacing each nested by its identity

The identity we use here is a plain string we compute by:

* serializing the original object
* computing an md5sum on the result

When we walk the object tree, each time we find a nested object for
which we already computed the identity, we replace it by its identity
value.

Finally, if we do not find any recursion, we just return the original
value we received. If there is a recursion, we cast the unnested array
to an object and give it to the asserter.