The basic resolver supports PHP7 aliases.
Since PHP7, some new keywords have been introduced, like `string` or
`float` for instance. Consequently, `asserters/string.php` have been
renamed to `asserters/phpString.php` etc.

However, it is like we forget to change the resolver too. This bug was
not visible because asserter aliases exist in the famous `test` class.
Nevertheless, if we use only the `asserter\generator`, the bug is
present.

This patch fixes the bug by checking twice if the class exists in the
resolver: First time with the canonical asserter name, second time
prefixed by `php` and with the first letter uppercased.