Avoid potential bug when mocking `preg_match`.
If a test case is mocking the `preg_match` function, it will create an
error in atoum's asserter resolver. Why? Because we are using
`preg_match` in `analyzer::isValidIdentifier` in `resolver::resolve`.

To avoid this issue (in atoum and in any potential test suites in the
wild), we bypass the function mocking system by using `\preg_match`
instead of `preg_match` (PHP will look directly at the root namespace,
which avoid us to be able to mock this function).