fix arguments priority parsing when thy have no priority
When an extension registered a new command line option,
and when the config file the loads this extension is loaded
via the command line, we could have issues depending on the order
of the arguments.

For example, this were not working:

```
 ./bin/atoum -c config/.atoum.php --autoloop
```

(we get this error `Error: Argument '--autoloop' is unknown, did you mean '--help'?`).

But this were working:

```
 ./bin/atoum --autoloop  -c config/.atoum.php
```

That because the --autoloop argument is not is the priorities array
beacause it's still not loaded.

So, in order to fix that, we always put the arguments with no priority
add then end.

Like that, other arguments, like adding a config file, that have the PHP_INT_MAX
value (the maximum priority), will be loaded before the arguments with no priority.
Then the extension will be loaded the the --autoloop argument will exist, and we will
not get an error.