Fix tests on test adapters
Since https://bugs.php.net/bug.php?id=69659, PHP handles calls to
__isset/offsetGet/offsetExists differently:

```php
<?php

class store {
    private $storage;

    public function __construct()
    {
        $this->storage = new ArrayObject();
    }

    public function __get($prop)
    {
        return $this->{$prop};
    }

    public function __isset($prop)
    {
        return false;
    }
}

$store = new store;
$store->storage[0] = uniqid();

var_dump(
    isset($store->storage),
    isset($store->storage[0]),
    isset($store->storage[1])
);
```

Before PHP 7.0.6, this code would return `false`, `true`, `false`. As of PHP 7.0.6, it returns
`false`, `false`, `false` because, before calling the `offsetExists` on the `ArrayAccess`, it calls
`__isset` on `store` which returns `false`, and it stops right away.

This is totally consistent and logical: why would PHP try to check an offset on something that is not
set? The `store` tells PHP the `storage` does not exist so it's likelly that the `storage[0]` and `storage[1]`
do not exist.