Ability to register additional error handlers for tests
This patch makes it possible to add one or several error handlers for a test.
This is particularly useful when one wants to deal with silenced errors.

One use case is to deal with Symfony `E_USER_DEPRECATED`, which are silenced on purpose.
Those errors are then caught by the debug bar.