Travis and HHVM: Fix various HHVM issues.
First, HHVM LTS is 3.12, so we stick on this version.
Second, there is a bug in HHVM where:

    $mixed instanceof \recursiveIterator

is falsy while

    $mixed instanceof \RecursiveIterator

is true. This is not a big chance for us but tests can be executed.

Third, this is probably another bug from HHVM (while I suspect the
specification to be unclear here) is the
`mageekguy\atoum\phing\task\path` constant that is defined to a mock
stream. An explicit cast to a string is required, else the constant is…
undefined! So:

    define('mageekguy\atoum\phing\task\path', (string) atoum\mock\streams\fs\file::get());