Report: Avoid reporting incorrect atoum path.
If the atoum is set to `/usr/local/atoum` for instance, and the
`$_SERVER['argv'][0]` is set to an absolute path like
`/usr/local/atoum/bin/atoum`, then the reported path will be
`/usr/local/atoum//usr/local/atoum/bin/atoum`, which is incorrect.

This patch computes the realpath of `$_SERVER['argv'][0]`, which
provides the expected output: The path to the atoum binary used to run
the tests.