feat: Avoid many allocations in error report.
When appending the stack trace of the error, the algorithm splits the
stack trace by line, and adds each line to the final error message.

In this situation, it is simpler to append the whole stack trace
directly without splitting it into an array, iterating over it, and
appending each item. The result is strictly the same but without memory
allocations and probably reallocations.